:truck:  rename: 각각의 step별 폴더로 정리
