+++ src/pages/CreatePostPage/components/FirstStep/Pin.tsx
