+++ src/pages/CreatePostPage/components/ThirdStep/ConfirmModal.tsx
@@ -1,6 +1,6 @@
-import { Modal, Button } from '../../../components/atoms';
-import { ButtonWrapper, ModalNoticeContainer, ModalNoticeText } from '../style';
-import { ConfirmModalProps } from '../types';
+import { Modal, Button } from '../../../../components/atoms';
+import { ButtonWrapper, ModalNoticeContainer, ModalNoticeText } from '../../style';
+import { ConfirmModalProps } from '../../types';
 
 const ConfirmModal = ({
   isModalOpen,
