+++ src/pages/CreatePostPage/components/ThirdStep/index.tsx
@@ -7,12 +7,12 @@ import {
   DateInput,
   GuideText,
   NextStepText,
-} from '../style';
-import { Text } from '../../../components/atoms';
+} from '../../style';
+import { Text } from '../../../../components/atoms';
 import { ChangeEvent, useEffect, useRef, useState } from 'react';
-import { ThirdStepProp } from '../types';
+import { ThirdStepProp } from '../../types';
 import ConfirmModal from './ConfirmModal';
-import UploadHeader from './UploadHeader';
+import UploadHeader from '../UploadHeader';
 
 const ThirdStep = ({
   latitude,
